+++ tests/lib/test_utils.py
@@ -1,6 +1,6 @@
-from binance.error import ParameterRequiredError
+from binance.error import ParameterRequiredError, ParameterTypeError
 from binance.error import ParameterValueError
-from binance.lib.utils import check_required_parameter
+from binance.lib.utils import check_required_parameter, check_type_parameter, convert_list_to_json_array
 from binance.lib.utils import check_required_parameters
 from binance.lib.utils import check_enum_parameter
 from binance.lib.utils import encoded_string
@@ -59,9 +59,25 @@ def test_fail_check_enum_parameter():
     ).should.throw(ParameterValueError)
 
 
+def test_pass_check_type_parameter():
+    value = []
+    check_type_parameter(value, "value", list)
+
+
+def test_fail_check_type_parameter():
+    value = []
+    check_type_parameter.when.called_with(
+        value, "value", dict
+    ).should.throw(ParameterTypeError)
+
+
 def test_encode_query_string():
     encoded_string({"foo": "bar", "foo2": "bar2"}).should.equal("foo=bar&foo2=bar2")
 
 
 def test_encode_query_without_email_symbol():
     encoded_string({"email": "alice@test.com"}).should.equal("email=alice@test.com")
+
+
+def test_convert_list_to_json_array():
+    convert_list_to_json_array(["symbol"]).should.equal("[\"symbol\"]")