+++ tests/spot/trade/test_my_trades.py
@@ -12,6 +12,7 @@ secret = random_str()
 
 params = {
     "symbol": "BTCUSDT",
+    "orderId": "345",
     "startTime": current_timestamp(),
     "endTime": current_timestamp(),
     "fromId": "1234567",