+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 -  `GET api/v3/exchangeInfo` now supports single or multi-symbol query
+-  `GET api/v3/myTrades` has a new optional field `orderId`
 
 ### Added
 - `GET /sapi/v1/c2c/orderMatch/listUserOrderHistory` to query user C2C trade history